Expose load_backends utilities
