+++ llm/backends/__init__.py
@@ -4,9 +4,8 @@ from __future__ import annotations
 
 from collections.abc import Callable
 from typing import Dict
-import importlib
-import importlib.metadata
-import pkgutil
+
+from .loader import discover_plugins, load_backends
 
 from .base import Backend
 from .superclaude import SuperClaudeBackend as _RealSuperClaudeBackend
@@ -32,6 +31,7 @@ __all__ = [
     "get_backend",
     "clear_registry",
     "discover_plugins",
+    "load_backends",
     "available_backends",
     "GeminiBackend",
     "OllamaBackend",
@@ -71,38 +71,4 @@ def available_backends() -> list[str]:
     return sorted(_BACKEND_REGISTRY)
 
 
-def discover_plugins() -> None:
-    """Import backend plugins so they register themselves."""
-    package = f"{__name__}.plugins"
-    paths: list[str] = []
-    try:
-        pkg = importlib.import_module(package)
-        paths = list(pkg.__path__)
-    except Exception:  # pragma: no cover - plugins package missing
-        pass
-
-    for mod in pkgutil.iter_modules(paths):
-        name = f"{package}.{mod.name}"
-        try:
-            module = importlib.import_module(name)
-        except Exception:  # pragma: no cover - optional dependency missing
-            continue
-        for attr in getattr(module, "__all__", []):
-            globals()[attr] = getattr(module, attr)
-            if attr not in __all__:
-                __all__.append(attr)
-
-    for entry in importlib.metadata.entry_points().select(group="llm.plugins"):
-        try:
-            module = entry.load()
-        except Exception:  # pragma: no cover - optional dependency missing
-            continue
-        for attr in getattr(module, "__all__", []):
-            globals()[attr] = getattr(module, attr)
-            if attr not in __all__:
-                __all__.append(attr)
-
-
-# Discover plugins at import time
-discover_plugins()
